Re-organize the setting view
