+++ app/build.gradle.kts
@@ -28,7 +28,7 @@ android {
         minSdk = 28
         targetSdk = 28
         versionCode = 2
-        versionName = "1.1.0"
+        versionName = "1.2"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
